Clean up SPI code
